+++ RayEngine/src/MovableCamera.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "cuda_runtime_api.h"
-#include "RayEngine/RayEngine.h"
+
 
 
 
@@ -11,17 +11,17 @@ class RAY_ENGINE_API RMovableCamera
 public:
 	RMovableCamera();
 	~RMovableCamera();
-	void strafe(const float& scale);
-	void set_fovx(const float& fovx);
-	void change_yaw(const float& scale);
-	void move_forward(const float& scale);
-	void change_pitch(const float& scale);
-	void rotate_right(const float& scale);
-	void change_radius(const float& scale);
-	void change_altitude(const float& scale);
-	void set_resolution(const float& x, const float& y);
-	void change_focial__distance(const float& scale);
-	void change_aperture_diameter(const float& scale);
+	void strafe(const double& scale);
+	void set_fovx(const double& fovx);
+	void change_yaw(const double& scale);
+	void move_forward(const double& scale);
+	void change_pitch(const double& scale);
+	void rotate_right(const double& scale);
+	void change_radius(const double& scale);
+	void change_altitude(const double& scale);
+	void set_resolution(const double& x, const double& y);
+	void change_focial__distance(const double& scale);
+	void change_aperture_diameter(const double& scale);
 
 
 	void build_camera(RCamera& camera);
@@ -32,11 +32,11 @@ public:
 
 	float3 position;
 	float3 view_direction;
-	float yaw;
-	float pitch;
-	float radius;
-	float apertude_radius;
-	float focial_distance;
+	double yaw;
+	double pitch;
+	double radius;
+	double apertude_radius;
+	double focial_distance;
 private:
